+++ component/loki/source/internal/kafkatarget/target_syncer_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/grafana/agent/component/loki/internal/fake"
+	"github.com/grafana/agent/component/common/loki/client/fake"
 
 	"github.com/grafana/dskit/flagext"
 	"github.com/prometheus/common/config"