+++ component/loki/source/journal/internal/target/journaltarget_test.go
@@ -13,7 +13,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/grafana/agent/component/loki/internal/fake"
+	"github.com/grafana/agent/component/common/loki/client/fake"
 
 	"github.com/coreos/go-systemd/sdjournal"
 	"github.com/go-kit/log"