+++ component/loki/source/syslog/internal/syslogtarget/syslogtarget_test.go
@@ -15,7 +15,7 @@ import (
 	"time"
 	"unicode/utf8"
 
-	"github.com/grafana/agent/component/loki/internal/fake"
+	"github.com/grafana/agent/component/common/loki/client/fake"
 
 	"github.com/go-kit/log"
 	"github.com/grafana/loki/clients/pkg/promtail/scrapeconfig"